Updates existing and add new 2nd level aliases
As part of the site re-org, we must add aliases to the make old links
work for users finding old references or using old bookmarks. For top
level targets that have moved into resources we add an alias to the old
location on the site. In the case of concepts, we have many nested pages
in sub-directories. These 2nd level links are also now accounted for.