Consolidation of some site information into resources section (#38)
* Initial page relocation for usnistgov/OSCAL-Pages#24.

* Remove header directives.

* Updates existing and add new 2nd level aliases

As part of the site re-org, we must add aliases to the make old links
work for users finding old references or using old bookmarks. For top
level targets that have moved into resources we add an alias to the old
location on the site. In the case of concepts, we have many nested pages
in sub-directories. These 2nd level links are also now accounted for.

* Add copy for how to use different resources.

* Update nav order with @Arminta-Jenkins-NIST's recommendations

Before the website reorganization on develop, the top-level navigation
had the following order.

- About
- Learn
- Concepts
- Reference
- Downloads
- Examples
- Tools
- Contribute
- Events
- Contact us

With previous commits and this final commit we have reorganized it like
so.

- News
- About
- Learn
- Resources
    - Concepts
    - Reference
    - Downloads
    - Examples
    - Tools
- Contribute
- Events
- Contact us

* Address alias indent error from @Arminta-Jenkins-NIST's feedback.